domain suggestions list added
