+++ lib/screens/profile/professional_info_add_screen.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import '../../widgets/p_i_a_s_widgets/p_i_a_s_chooseDomainField.dart';
 import 'package:provider/provider.dart';
 import 'package:flutter_icons/flutter_icons.dart';
 import '../../Model/doctor.dart';
@@ -7,7 +8,6 @@ import '../../providers/doctors_provider.dart';
 class ProfessionalInfoAddScreen extends StatelessWidget {
   static const routName = "profession-info";
   final _formKey = GlobalKey<FormState>();
-
   void _saveForm(BuildContext context, Doctor editedUser) async {
     if (_formKey.currentState.validate()) {
       _formKey.currentState.save();
@@ -34,7 +34,7 @@ class ProfessionalInfoAddScreen extends StatelessWidget {
           key: _formKey,
           child: Column(
             children: [
-              ChooseDomain(userData: _userData),
+              PIASChooseDomainField(userData: _userData),
               Container(
                 margin: EdgeInsets.symmetric(vertical: 10),
                 child: Row(
@@ -228,54 +228,6 @@ class ProfessionalInfoAddScreen extends StatelessWidget {
         onPressed: () {
           _saveForm(context, _userData);
         },
-      ),
-    );
-  }
-}
-
-class ChooseDomain extends StatelessWidget {
-  const ChooseDomain({
-    Key key,
-    @required Doctor userData,
-  })  : _userData = userData,
-        super(key: key);
-
-  final Doctor _userData;
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      margin: EdgeInsets.symmetric(vertical: 10),
-      child: Row(
-        children: [
-          Icon(
-            FlutterIcons.profile_ant,
-            size: 30,
-          ),
-          SizedBox(
-            width: 15,
-          ),
-          Expanded(
-            child: TextFormField(
-              initialValue:
-                  _userData.field != null ? _userData.field.toString() : "",
-              decoration: InputDecoration(
-                hintText: "e.g Psychiatrist",
-                labelText: "Domain",
-                border: OutlineInputBorder(
-                  borderRadius: BorderRadius.circular(5),
-                ),
-              ),
-              keyboardType: TextInputType.number,
-              onSaved: (value) {
-                if (value.isEmpty) {
-                  value = null;
-                }
-                _userData.field = value;
-              },
-            ),
-          )
-        ],
       ),
     );
   }
