+++ pubspec.yaml
@@ -43,6 +43,7 @@ dependencies:
   image_cropper: ^1.3.1
   firebase_storage: ^4.0.1
   flappy_search_bar: ^1.7.2
+  flutter_typeahead: ^1.8.8
 
 dev_dependencies:
   flutter_test:
